requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
#astronomer-cosmos
apache-airflow-providers-google
astronomer-cosmos
protobuf
paramiko
great-expectations
airflow-provider-great-expectations
apache-airflow-providers-dbt-cloud
apache-airflow-providers-sftp
#airflow.providers.dbt